requirements.txt
flask
joblib
spacy
streamlit
requests
gunicorn
mysql-connector-python
spacy>=3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
scikit-learn